chore(types): fix all 200 mypy-strict errors across the repo
Types-only cleanup — no runtime behavior changes. Brings `mypy .` from 200
errors to 0 (367 source files) so the mypy-strict CI check goes green.

Fixed across 28 files (annotations, return types, isinstance/None narrowing,
typing.cast, correctly-typed literals). 14 targeted `# type: ignore[code]`
remain, each with a reason, for genuine third-party/stdlib limitations:
- anthropic SDK's overloaded messages.create() passed through asyncio.to_thread
  (mypy can't resolve overloads via ParamSpec) — verifier.py, auto_label_examples.py
- cv2.VideoWriter_fourcc / cv2.normalize dst: present at runtime but missing/strict
  in cv2 stubs — video_maker.py, exclusive.py, recreate_session_videos.py, sam_detection.py
- assigning a MagicMock to bluesky.RunEngine in device tests — 4 test files

Verified: `mypy .` clean, `ruff check` + `ruff format` clean, and the test suite
shows the SAME results as the pre-cleanup baseline (zero regressions).

+++ scripts/projection_explorer.py
@@ -241,10 +241,10 @@ def compute_crop_bounds(
         return (0, volume.shape[1], 0, volume.shape[2])  # Full volume
 
     # Find center of mass and spread
-    cy = np.mean(y_coords)
-    cx = np.mean(x_coords)
-    y_std = max(np.std(y_coords), 20)  # Minimum spread
-    x_std = max(np.std(x_coords), 20)
+    cy = float(np.mean(y_coords))
+    cx = float(np.mean(x_coords))
+    y_std = max(float(np.std(y_coords)), 20.0)  # Minimum spread
+    x_std = max(float(np.std(x_coords)), 20.0)
 
     # Crop to sigma_mult * std around center
     y_min = int(max(0, cy - sigma_mult * y_std - padding))
